Move object queries and key into separate folder
